Simplify Authentication classes and make them copyable/movable (#425)
* Remove superfluous code

Simplifies storage of the Authentication and derived classes:

- remove username_ and password_ members because they are never used.
- take arguments as const lvalue reference rather than rvalues (since we
have no need to store them).
- remove pointless overrides in Digest/NTLM: the overridden destructor
does nothing, and the the GetAuthString() overrides simply call the base
class version (which is what happens anyway if they aren't overridden).
- Inherit Authentication's constructor in Digest/NTLM rather than
specifying an identical duplicate.
- remove digest.cpp and ntlm.cpp because they do nothing useful.

* Make Authentication and derived types movable/copyable

Because of the const auth_string_ the Authentication and derived classes
are neither copyable nor movable, which makes it more difficult to deal
with them.  Remove the const to allow it to be copied/moved and stored
in containers that require copyable and/or moveable elements.

* Avoid inherited constructors; change member to protected

* Add rvalue constructor

* Virtualize GetAuthString()

This allows the creation of custom subclasses that extend the class to
return an auth string in some other way.

* Simplify auth move ctor