KeyBlob introduced
Options let to set an SSL private key using a path to a file only.
Such approach prevents against setting the private key, which is stored
in a secure storage and is extracted using an algorithm.

Patch introduces KeyBlob, which translates to the CURLOPT_SSLKEY_BLOB directly.
It can be set using SslOptions::SetOption(const ssl::KeyBlob&) API.
It is assumed that blob content is PEM.