Debian files and scripts for debian package build
The library is packaged with the help of debmake and debuild.
This commit also updates the CI script to build packages on a debian
docker image with the help of the debmake and debuild packages.

The images that come out of this script are still not perfect!
lintian throws warnings and information messages that are still to be
investigated and solved.

To be tested.