Added subjectAltNames in certificates
This is required to be accepted as a valid certificate by the curl
commandline tool. The certificates are now issued for the localhost
addresses of IPv4 and IPv6 and for the DNS-name localhost.
This has worked in a local test.
The file paths in the tests still need to be adapted.