Rename private attributes in HttpsServer fixes bug
It is a convention to append a _ to private variables. Now, this is also
done in the HTTPS test server.
This renaming solves a name ambiguity issue in the HTTPS constructor.
Now, the attributes of the mg_tls_opts structs are filled with the
pointers to the sslCertFileName_ and sslKeyFileName_ strings which are
server attributes.

This could have been solved as well by pre-pending this-> to the
variables, but changing the naming convention seemed to be a cleaner
solution.