Fixed remaining failing session tests
The problem was that a few tests succeeded when executed independently,
but failed when executed together. The reason for this was that the
server method OnRequestLowSpeedTimeout executed its loop that sent
"Hello world!", waited and polled, even after the client disconnected.
Therefore, we need to introduce a check at the server whether the client
has disconnected and a wait at the client to be sure that the server
realizes that the client has disconnected.