Fixed failing error tests
The problem here was the nested call to mg_mgr_poll inside the server
response method. This method must be called in order to drain the header
of the message. However, if there are multiple connections at once, the
loop in mg_mgr_poll that goes through all connections which are stored
in a linked list causes problems, as the linked list gets inconsistent.
Therefore, a test calling the methods simulating slow connections needs
to wait before and after the request to make sure that only one
connection exists at a time.