AsyncWrapper: An extensible container for asynchronous responses
The new class template AsyncWrapper<typename T, bool> replaces std::future<T>
as the return value for async operations. It replicates std::future<T>'s
public member functions, so this change introduces no incompatibilities
with previous versions.

The only factor that changes is that the return value of async requests ceases
to have an is-a relation with std::future, which is a very minor change that
should introduce no incompatibilities, especially as the type alias
AsyncResponse is still in use, in the current iteration as an alias for
AsyncWrapper<Response>.

The goal of this change is for AsyncWrapper to support the cancellation of
operations while they are being processed.