Multiasync API: implementation of first operation and tests
Here, a possible implementation of an API that supports the compile-time
generation of multiple, asynchronous, cancellable requests is added. The
steps of the implementation are outlined below:

cpr::util::progressUserFunction: The original implementation is altered
in order to favour a function template. This has been judged necessary
because the function will now be able, through template instantiations,
to support any Callable target with the right signature, in this case
ProgressCallback and/or CancellationCallback. As added benefits,
template specializations will now not be instantiated if the code does
not use the function, and code locality may improve leading to marginal
performance gains.

cprtypes.h: Due to a change in libcurl's API, there are two possible
signatures for the progress function, depending on libcurl version.
Instead of the previous method of writing the possible signatures
explicitly and choosing via preprocessor macros, there is now a type
alias that gets the correct correspondence via macro.

cpr::Session: The function SetCancellationParameter receives a shared
pointer to an atomic boolean and uses it as cancellation parameter for
the transaction.

MultiAsync test family: Currently contains three tests, a basic multiple
request, a request cancellation while the requests are in the threadpool
queue, and a test that ensures that the cancellation function is called
with an acceptable frequency.