Multiasync Testing: Test all methods
The test suite MultiAsyncBasicTests performs multiple asynchronous
requests, using all implemented Http methods, and evaluates that results
are correct. The requests that have been combined come from the test
suites of the appropriate method.

A missing template parameter that does not get deduced and broke the
tests for ubuntu 18 (gcc 7) has been fixed.

A possible source of a non-deterministic bug that caused the final test
(interval to complete N calls to the progress function) to occasionally
fail with a mutex-ownership-related exception has been fixed.