Update CMakeLists.txt
I only get this error when compiling with ndk. Linux has no problem, it compiles on Linux with or without the -lstdc++fs flag, but ndk cannot find such a library, but you can use the filesystem library without this flag (-lstdc+++fs).


This change is necessary to improve compatibility and usability of the cpr library. Thank you for your attention to this issue.