Remove move semantics from Authentication
move semantics with a std::string introduce difficulty with secure
erasing because we have no guarantee that the moved-from object doesn't
still have the the credentials (because of small string optimization).

This removes the move ctor and assignment so that we always get secure
erasing on destruction without any weird SSO cases to worry about.