Find the path to the certificate bundle in runtime
The change makes the compiled library work on
different distributions, even if the path to the
certificate bundle in runtime is different compared
to the path where it was during compilation.

This can be useful when the compilation is done on
a manylinux Docker image which is based on CentOS,
but the compiled library/binaries are used on
Ubuntu or Debian.