add curl's ANY and ANYSAFE authentication options
ANY lets curl determine the correct authentication method. ANYSAFE is
similar but it eliminates the 'BASIC' option.

Some SOAP servers do not reply to authentication requests if content
of the initial request is empty. When ANY option is selected curl
sends initial request with full contents, which allows authentication
process to work.