Remove an early return in ~MultiPerform
This curl_multi_remove_handle call is a bit more error-prone with the
changes we've made, and skipping over sessions in the cleanup process
leads to sessions that can't be used due to isUsedInMultiPerform being
set to true, along with other potential, undesired consequences as a
result of improper cleanup.

We could remove the curl_multi_remove_handle call entirely, but that,
too, will cause issues if any part of the code throws after we've added
handles to the multi stack and before removing them.

It's not something that's very likely to happen with the existing code,
but that could easily change in the future and the call to that function
doesn't seem to be doing any harm, aside from some wasted CPU cycles.