test: don't reflect Content-Length from request
Previously, POSTing to this endpoint would mean that the response would
have two Content-Length headers: the one reflected from the request, and
the new one that is for the response. This is improper, so this commit
adds Content-Length to the headers excluded from reflection.

This was causing curl (as of 8.17.0, though the change might have
happened earlier) to give a "weird server reply" error, which made
several tests fail.